Added basic html and css
