+++ app/templates/hobbies.html
@@ -1,5 +1,27 @@
 {% extends 'base.html' %}
 <!-- Add content here -->
 {% block content %}
-<h1>Hobbies</h1>
+    <!doctype html>
+    <html>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <script src="https://cdn.tailwindcss.com"></script>
+    </head>
+    <body>
+        <div class="mt-4 mb-[20rem] flex justify-center">
+        <div class="grid grid-cols-1 gap-02 md:grid-cols-2">
+            {% for hobby in hobbies %}
+                <div class="transition-[max-height] ease-in-out duration-500 m-6 max-w-xs bg-white rounded-lg max-h-[13rem] hover:max-h-[40rem] overflow-hidden border border-gray-200 shadow-md dark:bg-gray-800 dark:border-gray-700">
+                    <img class="rounded-t-lg" src="{{ hobby.img }}" alt="" />
+                    <div class="p-6">
+                    <h5 class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">{{ hobby.name }}</h5>
+                    <p class="mb-3 font-normal text-gray-700 dark:text-gray-400">{{ hobby.desc }}</p>
+                    </div>
+                </div>
+            {% endfor %}
+        </div>
+        </div>
+    </body>
+    </html>
 {% endblock %}
